1. 그룹 페이지 동작 활성화 2. 로그인 시 온보딩 페이지로 이동하는 로직 생성

+++ app/src/main/java/com/wewrite/android/api/repository/GroupRepository.kt
@@ -35,7 +35,7 @@ class GroupRepository(private val groupService: GroupService) {
         return groupService.joinGroup(request)
     }
 
-    suspend fun leaveGroup(groupId: Int): BaseResponse {
+    suspend fun leaveGroup(groupId: Long): BaseResponse {
         return groupService.leaveGroup(groupId)
     }
